requirements.txt
django==4.0.4
djangorestframework==3.13.1
Pillow==9.1.1
django-rest-knox==4.2.0g